pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ronin-supply"
version = "0.1.0"
description = "AI-powered Retail Operating System with hyper-local intelligence"
authors = [
    {name = "Ronin Supply Team", email = "team@roninsupply.com"},
]
dependencies = [
    # Core FastAPI and async support
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    
    # Database and storage
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.23",
    "alembic>=1.13.0",
    "supabase>=2.0.0",
    "redis>=5.0.0",
    
    # AI and ML services
    "google-generativeai>=0.3.0",
    "openai>=1.3.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.40",
    "nixtla>=0.5.0",
    
    # Web intelligence and APIs
    "tavily-python>=0.3.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    
    # Geographic and spatial data
    "geoalchemy2>=0.14.0",
    "shapely>=2.0.0",
    "geopy>=2.4.0",
    
    # Data processing and validation
    "pydantic>=2.5.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "pillow>=10.1.0",
    
    # Email and communication
    "aiosmtplib>=3.0.0",
    "email-validator>=2.1.0",
    
    # Environment and configuration
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.1.0",
    
    # Monitoring and logging
    "structlog>=23.2.0",
    "prometheus-client>=0.19.0",
    
    # Development and testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.92.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true